ADD - 도서 목록 API 작업중

+++ src/main/kotlin/com/group/libraryapp/security/AuthenticationDTO.kt
@@ -3,13 +3,13 @@ package com.group.libraryapp.security
 import com.group.libraryapp.domain.user.UserType
 
 class AuthenticationDTO(
-    val id: Int,
+    val id: Long,
     val email: String,
     val name: String,
     val userType: UserType
 ) {
     companion object {
-        fun of(id: Int,
+        fun of(id: Long,
                email: String,
                name: String,
                userType: UserType = UserType.USER): AuthenticationDTO {
